@@ -1,37 +1,53 @@
 import React, { Component } from 'react'
 import { Route, Switch } from 'react-router-dom'
-import { currentCreditAPI } from './api/BankApi'
 import AddCreditPage from 'src/components/pages/AddCreditPage'
 import HomePage from 'src/components/pages/HomePage'
 import SearchResultsPage from 'src/components/pages/SearchResultsPage'
 import HomeDetailPage from './components/pages/HomeDetailPage'
+import AddHomePage from './components/pages/AddHomePage'
 
 export default class App extends Component {
   constructor (props) {
     super(props)
     this.state = {
-      customerCredit: '0',
+      user: null
     }
   }
 
-  componentWillMount() {
-    currentCreditAPI().then((response) => {
-      this.setState({customerCredit: response})
-    })
+  componentDidMount() {
+    if (localStorage.hasOwnProperty('user')) {
+      let receivedObj = localStorage.getItem('user')
+      this.setState({user: JSON.parse(receivedObj)})
+    }
   }
 
   _handleCreditChange = (amount) => {
+    const { user } = this.state
     if (!isNaN(amount)) {
-      const newCredit = parseInt(this.state.customerCredit) + parseInt(amount)
-      this.setState({customerCredit: newCredit})
+      const newCredit = parseInt(user['credit']) + parseInt(amount)
+      let newUser = user
+      newUser['credit'] = newCredit
+      this.setState({user: newUser})
     }
+  }
+
+  _handleLogin = (loggedInUser) => {
+    this.setState({user: loggedInUser})
+    localStorage.setItem('user', JSON.stringify(loggedInUser))
+  }
+
+  _handleLogout = () => {
+    this.setState({user: null})
+    localStorage.clear()
   }
 
   _renderHomePage = () => {
     return (
       <HomePage
-        credit={this.state.customerCredit}
+        user={this.state.user}
         onCreditChange={this._handleCreditChange}
+        onLogin={this._handleLogin}
+        onLogout={this._handleLogout}
       />
     )
   }
@@ -39,8 +55,10 @@ export default class App extends Component {
   _renderAddCreditPage = () => {
     return (
       <AddCreditPage
-        credit={this.state.customerCredit}
+        user={this.state.user}
         onCreditChange={this._handleCreditChange}
+        onLogin={this._handleLogin}
+        onLogout={this._handleLogout}
       />
     )
   }
@@ -49,8 +67,10 @@ export default class App extends Component {
     return (
       <SearchResultsPage
         match={props.match}
-        credit={this.state.customerCredit}
+        user={this.state.user}
         onCreditChange={this._handleCreditChange}
+        onLogin={this._handleLogin}
+        onLogout={this._handleLogout}
       />
     )
   }
@@ -59,8 +79,22 @@ export default class App extends Component {
     return (
       <HomeDetailPage
         match={props.match}
-        credit={this.state.customerCredit}
+        user={this.state.user}
         onCreditChange={this._handleCreditChange}
+        onLogin={this._handleLogin}
+        onLogout={this._handleLogout}
+      />
+    )
+  }
+
+  _renderAddHomePage = (props) => {
+    return (
+      <AddHomePage
+        match={props.match}
+        user={this.state.user}
+        onCreditChange={this._handleCreditChange}
+        onLogin={this._handleLogin}
+        onLogout={this._handleLogout}
       />
     )
   }
@@ -83,6 +117,10 @@ export default class App extends Component {
         <Route
           exact path='/HomeDetail/:houseId'
           render={this._renderHomeDetailPage}
+        />
+        <Route
+          exact path='/AddHome'
+          render={this._renderAddHomePage}
         />
       </Switch>
     )
