@@ -6,10 +6,10 @@ import 'src/styles/General.css'
 
 export default class GeneralHeader extends Component {
   render() {
-    const { pageTitle, credit } = this.props
+    const { pageTitle, user, onLoginModalOpen, onLogout } = this.props
     return (
       <header className="addCreditHeaderContainer defaultHeaderHeight rtl">
-        <HeaderUpperRow credit={credit}/>
+        <HeaderUpperRow user={user} onLoginModalOpen={onLoginModalOpen} onLogout={onLogout} />
         <HeaderImageBar pageTitle={pageTitle}/>
       </header>
     )
