@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import { Redirect } from 'react-router-dom'
-import { searchParamsAreValid } from 'src/Utilities/formats'
+import { searchParamsAreValid, toEnglish } from 'src/utilities/formats'
 import { messages } from 'src/constants/FaTexts'
 import SearchFormUpper from './SearchFormUpper'
 import SearchFormLower from './SearchFormLower'
@@ -8,9 +8,7 @@ import 'src/styles/HomePage/HomePageHeader.css'
 import 'src/styles/HomePage/HomePageResponsive.css'
 import 'src/styles/SearchForm/SearchForm.css'
 import 'src/styles/SearchForm/SearchFormMediaQuery.css'
-import { toEnglish } from 'src/Utilities/formats'
 import Fa from 'src/constants/Fa'
-
 
 
 export default class SearchForm extends Component {
@@ -50,7 +48,6 @@ export default class SearchForm extends Component {
   }
 
   _createUrl = (maxPrice, minArea, propertyType, dealType) => {
-    // let persianDealType = (dealType === 'rental') ? Fa["rent"] : Fa["buy"]
     return  '/SearchResults' +
             '/' + (maxPrice || 'none') +
             '/' + (minArea || 'none') +
@@ -68,22 +65,18 @@ export default class SearchForm extends Component {
 
   _handleMaxPriceChange = (newMaxPrice) => {
     this.setState({maxPrice: newMaxPrice})
-    // console.log("NEW MAX PRICE ====> " + newMaxPrice)
   }
 
   _handleMinAreaChange = (newMinArea) => {
     this.setState({minArea: newMinArea})
-    // console.log("NEW MIN AREA ====> " + newMinArea)
   }
 
   _handlePropertyTypeChange = (newPropertyType) => {
     this.setState({propertyType: newPropertyType})
-    // console.log("NEW PROPERTY TYPE ====> " + newPropertyType)
   }
 
   _handleDealTypeChange = (newDealType) => {
     this.setState({dealType: newDealType})
-    // console.log("NEW DEAL TYPE ====> " + newDealType)
   }
 
   _handleSearchClick = (event) => {
