@@ -2,22 +2,44 @@ import React, { Component } from 'react'
 import Layout from 'src/components/general/Layout/Layout'
 import Fa from 'src/constants/Fa'
 import AddCredit from 'src/components/general/AddCredit/AddCredit'
+import LoginDialog from '../general/Login/LoginDialog'
 
 
 export default class AddCreditPage extends Component {
+  constructor (props) {
+    super(props)
+    this.state = {
+      modalOpen: false,
+    }
+  }
 
+  handleModalClose = () => {
+    this.setState({ modalOpen: false });
+  }
+
+  handleModalOpen = () => {
+    this.setState({ modalOpen: true });
+  }
 
   render () {
+    const { user, onCreditChange, onLogin, onLogout } = this.props
     return (
       <Layout
         isHomePage={false}
         pageTitle={Fa['add credit page']}
-        credit={this.props.credit}
-        onCreditChange={this.props.onCreditChange}
+        user={user}
+        onCreditChange={onCreditChange}
+        onLoginModalOpen={this.handleModalOpen}
+        onLogout={onLogout}
       >
+        <LoginDialog
+          open={this.state.modalOpen}
+          onDialogClose={this.handleModalClose}
+          onLogin={onLogin}
+        />
         <AddCredit
-          credit={this.props.credit}
-          onCreditChange={this.props.onCreditChange}
+          user={user}
+          onCreditChange={onCreditChange}
         />
       </Layout>
     )
